misc (make, test): add tests and restructure Makefile
Able to build unit-tests now.
Able to

Add header and source files for Test.
Add memory-size with ERASE macro.
Add fill memory with Zeroes while erasing them.
Add `make again` in Makefile for clean build.

Resolves: #2

+++ include/node.h
@@ -29,7 +29,7 @@ typedef struct NODE {
 Node* create_node (Node_Type type);
 Node* duplicate_node (Node* node);
 void delete_node (Node** node_address);
-void set_node_name (Node* node, int name_length, char* name_address);
+void set_node_name (Node* node, size_t name_length, char* name_address);
 void display_node (Node* node);
 void display_special_node (Node* node);
 void display_node_details (Node* node);
